mh/requirements.txt
selenium>=4.10.0
Pillow>=9.5.0
